Dockerfile updated to run sql server and also controller and model discrepencies resolved

+++ Services/Order/Order.Application/Responses/OrderResponse.cs
@@ -3,7 +3,7 @@
 public sealed record OrderResponse(
     string Id,
     string UserName,
-    decimal TotalPrice,
+    double TotalPrice,
     string FirstName,
     string LastName,
     string EmailAddress,
